Improved life cycle logic
