+++ Runtime/Constants.cs
@@ -1,8 +1,10 @@
 namespace Padoru.Core
 {
-    public class Constants
+    public static class Constants
     {
         public const string SETTINGS_OBJECT_NAME = "PadoruSettings";
+        
+        public const string PROJECT_CONTEXT_TAG = "ProjectContext";
         
         public const string GAME_FSM_TAG = "GameFSM";
