+++ Runtime/Initialization/PadoruApplication.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 053587f5149fcda4cb082c842f2ab047
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
