+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "com.padoru.core-lib",
-  "version": "2.1.0",
+  "version": "2.2.0",
   "displayName": "Padoru Core",
   "description": "A core library to develop games in Unity",
   "unity": "2020.3",
